gatekeeper: add setup, Makefile, sample
Add a setup script to configure and build DPDK and a Makefile to build
the gatekeeper application. Also add the starter file for gatekeeper,
which currently just initializes the EAL and exits to show it can use
the DPDK libraries.

The README gives instructions on how to build DPDK and gatekeeper and
configure the environment for hugepages and network adapters.