gatekeeper: add functional blocks, lib, include
Add the skeleton code for each of the functional blocks, as well as a
subdirectory for library functions that may be used by multiple blocks.
Also add a directory for all includes that will be needed by the main
Gatekeeper code. Adjust the Makefile to include all these files in the
build process.