setup.sh: make setup re-runnable
The setup script previously continued to add module names to
/etc/modules every time it was run. Particularly if there are
multiple users on one system, this can clutter /etc/modules.
Now, they are only added if they are not already present.

Other system-wide alterations, such as loading modules and adding
a symlink, can (and should) be attempted multiple times. If a module
is already loaded, there's no harm in trying to load it again. If a
symlink already exists, the ``ln'' command outputs an error to the
console to warn the user.