config: push block launching to static config
The goal is for the static configuration component to read in
configuration information for the needed blocks and then run
them using separate lcores. This commit adds configuration structs
for each block and moves the launching to the static component.

If running any of the blocks fail to launch, the exiting variable
is set to true so that each block can stop and clean up before
returning.