net: improve queue configuration
This patch removes the assumption that all blocks use both the TX
and RX queues on both interfaces. It allows each block to be assigned
a queue identifier on a per-interface, per-queue type (RX or TX)
basis. Gatekeeper now permits different numbers of RX queues and TX
queues to be created on the different interfaces.

To do so, the configuration must be split up. This is because these
functions must be called in this order:

 - rte_eth_dev_configure()
 - rte_eth_rx_queue_setup()/rte_eth_tx_queue_setup()
 - rte_eth_dev_start()

Therefore, the net configuration code must configure each port,
and then allow the functional blocks to run so that they may
call the rte_eth_{r/t}x_queue_setup() functions via get_queue_id().
Then, the functional blocks wait for the devices to be started.