net: add interface abstraction
This patch creates an identifier for both the front and back
interfaces so that all per-port operations can be done on
one identifier. The operator configures the ports in
lua/net.lua by identifying which ports belong to which interface
using the (kernel interface) name of the port. The configuration
then sets up all needed ports.

In order for the operator to be able to use the interface name,
the setup.sh script saves the interface names in a map.

To map a PCI address to a port, a library function needed to be
exposed from the DPDK library, so the submodule is updated.