lls: add initialization, cache, and ARP support
To keep track of neighbor data, LLS uses a cache implemented using
the hash library. Functional blocks should express interest in an
entry by supplying a callback function that is invoked when the
entry is available or when it is updated.

Add ARP to be able to resolve IPv4 addresses to Ethernet addresses.
To do so, the LLS block adds L2 filters for ARP on both front and back
interfaces and processes ARP packets using an address resolution cache.

Also, an interface is exposed for functional blocks to hold and put
ARP maps.