net: replace RTE_VERIFY() with proper error handling
The problem is that the test in a RTE_VERIFY() was requiring that
there were the exact number of interfaces available to DPDK as
the number of interfaces that Gatekeeper is going to use.
This condition is not always true since one ore more interfaces
can be used by another application such as pktgen.

Moreover, some of the conditions of the test in RTE_VERIFY() could
happen in practice without being a bug.
For example, the configuration could require more network ports than
Gatekeeper can handle.

This patch replace the RTE_VERIFY() with proper error handling.

This issue has been ignore because there was a RTE_ASSERT()
instead of RTE_VERIFY() before, so no checks was being done.