lls: add ND for IPv6 address resolution
Add the ND protocol to the LLS block to resolve IPv6 addresses to
Ethernet addresses.

The ND implementation follows RFC 4861, except where noted in
the file lls/nd.c. To do this, the NIC needs to be configured to
receive IPv6 multicast Ethernet frames, and the network library
needs to be changed to include a link-local IPv6 address and
solicited-node multicast addresses.

The LLS block currently cannot directly receive ND packets. Therefore,
the RSS receivers on both interfaces must classify and distribute ND
packets to the LLS block. On the front interface, the GK and GT blocks
do this (depending on whether we're running Gatekeeper or Grantor).
There is currently no RSS receiver on the back interface, so this patch
enables the LLS block to be the RSS receiver on the back interface until
a different block can do it. At that point, that block should similarly
send the LLS block all ND packets.

If we continue to steer IPv6 packets in software, we should make use
of the DPDK ACL/classification and distribution libraries.

To allow an IPv6 multicast Ethernet address to be assigned to a bonded
interface, add the rte_eth_dev_set_mc_addr_list() implementation for the
bonding driver. This makes sure the bonded driver has all the
functionality needed for IPv6 multicast Ethernet frames. Also update
the dependencies/dpdk folder as a whole.