net: add LACP to net config and LLS
Enable the Link Aggregation Control Protocol (LACP) to automatically
bond physical ports to a single logical channel.

DPDK supports a bonded driver library that runs LACP without explicit
calls from the application. Every time the application calls the RX/TX
functions on a port, LACP messages are received or injected as appropriate
to negotiate the bundling of links with a peer.

The only requirements to make this work are to ensure that the RX/TX
functions are called at least every 100 ms and that the number of
buffers supplied to the RX function is at least twice the number of
slaves. This allows LACP frames to be sent and received within the
timing expectations of the protocol.