gk/gt: fix configuration cleanup
The cleanup for the GK block if configuration fails is broken. The
reference count is set to 0 by rte_atomic32_init() in run_gk(),
but gk_stage1() tries to decrement that reference count when
setting up a GK instance fails. That would cause the reference count
to go to -1, and cleanup_gk() would not be invoked.

Also, other failure conditions in gk_stage1() and gk_stage2() were
not doing any cleanup at all. Same for the GT block.

This patch removes reference counting during the configuration stages
of the GK block, and makes both the GK and GT blocks cleanup during
failure in stage 1 and stage 2. It also fixes a style issue: when
testing for non-NULL, the comparison must be explict in cleanup_gk().