gk: add functionalities to edit LPM table
This patch mainly adds support for addition/deletion of FIB entries
for the global GK LPM table, e.g., add_fib_entry() for addition, and
del_fib_entry() for deletion.

In general, a Grantor FIB entry will be added if packets for a
destination prefix are forwarded to a Grantor server; a Gateway FIB
entry will be added if packets for a destination prefix are forwarded
to either front or back network; a drop FIB entry will be added if
packets for a destination prefix need to be dropped.

For the deletion of a FIB entry, before we remove it,
we need to notify GK blocks of all removals of FIB entries.
After all the GK blocks complete their corresponding process, then we
can safely remove the FIB entry.