lls: minor code touch-ups
Since submit_nd() is only used within the LLS block, make it static.

When a recent patch moved the ND packet length check inside of
process_nd(), it did not also move the initializations of the
packet header pointers to after the length has been verified.
This patch makes sure the relevant parts of the packet exist
before obtaining pointers to that part of memory.