cps: configure KNIs using libmnl
Add functionality to assign IP addresses to the KNIs that mirror
the IP addresses of the Gatekeeper interface, and enable those
interfaces to be brought up. The structure of using modify_ipaddr()
and modify_link() to pass different flags should make it easier
for the dynamic configuration in the future to use the same code
to delete IP addresses, replace IP addresses, bring a KNI interface
down, etc.