cps: add address resolution to KNI
The KNI is expected to at least send control plane packets (IPv4
and IPv6 BGP) as well as resolution packets for ARP and ND. This
patch adds code to catch ARP and ND packets and handle them. All
other packets, including BGP packets, are forwarded to the
appropriate Gatekeeper interface.