gk: integrate LLS to the GK blocks
Since the Ethernet header cache can be modified
by LLS block, and the GK blocks need to read them
to decide the nexthop's link layer header, this patch
adds support for the sequential lock to deal with
the concurrency.

In addition, some extra modification needed to the patch
that edits the GK LPM table, since the operations on
GK LPM table will affect the Ethernet header cache.