acl: allow software ntuple filtering
If ntuple filters are not supported, then the code will notify the
user and automatically transmit IPv4 BGP and GK-GT Unit packets
received in the GK and GT blocks to the CPS and GGU blocks respectively
for further processing.

Some code used to process ACLs is the same between IPv4 and IPv6,
and can be factored out to reduce code duplication. The code that
can't be factored out includes definitions of IPv4/IPv6 rule types
that are specific to each protocol.