gk/ggu/gt: add support for VLAN
In general, this patch makes the GK, GGU, and GT blocks aware that the
L2 header can be of variable size: specifically, there can be a four
byte VLAN header.

Whenever a packet is received by one of these blocks, if it has
a VLAN header it is marked while fetching the next header, using
the pkt_in_skip_l2() function. This makes it easier to later adjust
the size of the packet buffer, if needed.

There are special cases when the blocks need to be aware of the
presence of a VLAN header, and for efficiency purposes alter their
behavior so that they are not changing the packet buffers multiple
times. For example, when packet buffers are re-used, their L2
header length may need to be adjusted using a function like
adjust_pkt_len(). Other times, the packet buffer is already being
adjusted (like in encapsulate() and decap_and_fill_eth()); in these cases,
more logic is added to account for how variations in the L2 header
length affect the number of bytes added and removed.