lls/cps/acl: add support for VLAN
In general, this patch makes the LLS and CPS blocks aware of the
variability in L2 header lengths.

The KNI is kept unaware of any VLAN headers: we remove any VLAN headers
when they are on packets bound for the KNI, and insert them into any
packets coming from the KNI when configured to do so.

To be able to match ND and IPv6 BGP packets, the ACL library also
needs to be aware that the filters should be applied starting after
a variable length L2 header.