acl: use new constant for no ACL match
The DPDK ACL library returns a value of 0 for packets that
do not have a match in the ACL. Previously, the ACL library
did not allow 0 to also represent a match that should be
treated as if it were not a match, but DPDK now permits the
use of 0 for this purpose. We don't use 0 for any additional
meaning, but we need to use our own constant to represent
to represent no match.