gt/ggu: update GT-GK Unit packet format
The format for packets sent from the GT block to the GT-GK Unit
is changed to be able to mix different policy decisions instead
of requiring them to be sequentially ordered by decision type.
This makes future patches for batching policy decisions easier.

This sets up a resolution for issue #58.