gk: fix bugs and style issues
This patch fix the issues related to the function calls:
rte_lpm_is_rule_present() and rte_lpm6_is_rule_present().

The above two functions will return 1 if the rule exists,
0 if it does not, a negative value on failure.

Also, this patch add a newline for all the log messages.
Otherwise, the log message is hard to parse.