gk: fix FIB deletion bugs
This patch closes #111. The root cause of this bug was:
the parameter 'struct ipaddr *addr' in ether_cache_put()
is from the Ethernet cache entry. After calling put_xxx(),
it's possible that gk_arp_and_nd_req_cb() is called before
rte_hash_del_key(). In this case, the 'addr' will be reset,
so that the hash key (i.e., addr->ip.v4 or addr->ip.v6) becomes invalid.

The solution is to make a copy of the hash key, so that
even the cache entry is reset, we can still remove the key-value
pair from the hash table.