gt: batch GGU policy decision notification packets
Batch policy decision notifications to reduce the number of
packets that the GT-GK Unit has to handle.

To do so, maintain a fixed array of notification packets that
is periodically flushed. When a new Gatekeeper server is
encountered and we need to make room for its notification packet,
send a random packet.

This resolves #58.