net: add flag for iface initialized
Gatekeeper interface initialization can fail and be cleaned
up at multiple points. When Gatekeeper closes, the network
is also cleaned up. Therefore, during failure of initializing
the interfaces, multiple cleanups were attempted which led
to segmentation faults. Now when Gatekeeper interfaces are
initialized anytime after the Lua initialization, the alive
flag is set so that we can mark whether to free it.