net: enable Gatekeeper to be run without RSS
To allow Gatekeeper to run without RSS, we disable RSS by default
and only turn it on if the interface supports it.

When running without RSS, the delivery of packets to queues is
dependent on the underlying device and driver. Drivers like
virtio distribute packets using RFS (to all queues). Other
devices just deliver all packets to queue 0. Therefore, when
RSS is disabled we use only one RX queue, which is used by either
a GK or a GT block. This additionally requires only one GK or GT
instance to be running when RSS is disabled.

RX queues for all other blocks are only used when RSS is supported.
This is because the other blocks tell the network configuration
they need queues before it can be determined whether filters are
supported, and once the network configuration specifies a number
of queues for the port, all of those queues have to be setup or
else a segmentation fault occurs in DPDK. Therefore, the CPS,
GGU, and LLS blocks request to use an RX queue and configure it
if RSS is available, but only actually use it if both RSS and
their type of filter is available. If RSS is disabled, then
the network configuration can figure out to setup only one
queue, and therefore the other blocks do not need to setup the
queue they requested.

This resolves #92.