cps: add dynamic logging type
This patch adds a logging type for the CPS block by following
the example of various DPDK libraries (such as lib/librte_ethdev).
This involves registering for a dynamic logging type and replacing
RTE_LOG() with a block-specific macro, CPS_LOG().

It also adds the ability to change the logging level of the
CPS block while Gatekeeper is running via the dynamic configuration.