lls: adopt struct ipaddr in LLS cache
Adopt struct ipaddr to use as a key for the LLS caches, which
allows us to use more general functions and remove IPv4 and
IPv6-specific functions from the cache library.

To help do this, the patch introduces MAX_INET_STRADDRLEN to
represent the maximum length of an IP address string.