lua: match example FIB entries with net config
When adding entries to the FIB, the gateway should be in
the same subnet as the IP address of the Gatekeeper server,
and the route prefixes should not contain the IP addresses
of the Gatekeeper server as well. This patch makes the
example FIB entries match the example network configuration.

This closes #173.