cps: add option to read from other routing daemon
Add the configuration option to bind the routing daemon Netlink
socket to a specific port, enabling a routing daemon other than
the kernel to send Gatekeeper Netlink messages directly. Update
the names of the relevant functions and variables to emphasize
that these Netlink interactions are with a routing daemon (rd).

Also, add BIRD as a submodule dependency.

This patch is related to #56.