cps: keep track of KNI index
When sending routes to the routing daemon during table scans,
Gatekeeper needs to include information about interfaces
associated with the routes. To avoid having to do a Netlink
lookup of the kernel-assigned interface index with each scan,
we look up the index of each KNI once and save it in Gatekeeper.

This patch is related to #56.