net: verify offload capabilities of ports
Each port should be checked to make sure it supports
RSS functions and other offload capabilities before the
ports are initialized.

Ensuring RSS and offload capabilities before configuring
the ports also ensures that when there are multiple ports
in a Gatekeeper interface and only some support RSS, a
consistent configuration is constructed.

This patch closes #184.