net: add additional checks for RSS hashes
Split up ETH_RSS_IP into IPv4-related and IPv6-related hash functions.
For each type of IP being used in Gatekeeper, check the supported
hashes of the device. If none are supported, disable RSS.

Use ETH_RSS_IPV4|ETH_RSS_IPV6 as a more specific expectation for
the hashes that the device should support, and issue a warning
if they don't for each version of IP that is enabled.

This patch closes #152.