net: apply filters to individual ports
EtherType and ntuple filters were being applied to the interface
ID instead of individual ports.

This worked for Gatekeeper interfaces with only a single port,
since the interface ID will match the port ID in that case. But
when there are multiple ports, the interface ID identifies the
bonded port, which does not support these filters.

To use these filters when multiple ports are configured, this
patch applies the filters to the ports directly instead of the
Gatekeeper interface IDs.

A filter is only considered supported by a Gatekeeper interface
if all component ports in the interface support it.