cps: enable multiple NL messages in each packet
When the CPS block receives Netlink messages from a routing
daemon, allow the possibility of multiple messages per packet.

Since we do not know how many messages there are per packet
a priori, we limit the number of update packets instead of
the number of update messages. Once we do that, we no longer
have to track a series of updates and pass them back and
forth; we can just process each message one at a time and
add and delete routes as necessary.