cps: add NLMSG_DONE to batch
When we add an NLMSG_DONE message to the end of a batch, we need
to invoke mnl_nlmsg_batch_next() to make sure the batch data
structure is updated with the new length. Otherwise, the DONE
message won't be sent.

This bug did not appear previously because we reply with an ACK
to the requesting daemon, but the DONE message is needed in the
general case when using NLM_F_MULTI messages.