ggu: remove block name from variables
Including ggu in variable names for this block is redundant
and unnecessary. Variables ggu_src_port and ggu_dst_port
are exceptions because they are referring to the GGU protocol
used between Grantor blocks and the GGU block, and also match
variables of the same name in the GT block.