gt: allocate Lua states on NUMA nodes
This patch almost closes issue #143, but it doesn't do so because
the current version of DPDK that we are using does not include
rte_realloc_socket().

Once we upgrade our DPDK version, one must replace rte_realloc()
with rte_realloc_socket() in alloc_lua_mem_in_dpdk() to close
issue #143.