gk: adjust type of granted state of a flow entry
This patch changes the type of the field tx_rate_kb_cycle of
the granted state of a flow entry from int to uint32_t in order to
match the parameter tx_rate_kb_sec that comes from policies.
The type of the field budget_byte is also upgraded from int to
uint64_t in order to guarantee that the product
tx_rate_kb_cycle * 1024 is always in range.