cps: add fixes to CPS block
In order for the KNI carrier state to be set as up, the function
rte_kni_update_link() needs to be called. This allows programs like
BIRD to recognize the interface as up using the IFF_LOWER_UP flag.

Also, the MAC address for the KNI needs to be explicitly set if
we want it to match the MAC address of the Gatekeeper interface.