gk: scan over flow table entries directly
Instead of iterating over the hash table map, iterate over
the flow entries directly.

To do so, add a field to each flow entry which marks whether
that entry is actively in use. To keep the size of the flow
entry <= 128 bytes, change the state to be a uint8_t.