gk: drop new flows over batches when flow table is full
This patch solves the performance degradation issue https://github.com/AltraMayor/gatekeeper/pull/330.
The key insight is that having a single failure per batch
may be bearable. Given a batch of received packets,
it tries to add all new flows, but if one addition fails
for the batch of received packets, it doesn't try all
the other flows to be added in that batch.