sol: solve the contention issue
The function call common_ring_mc_queue() has very high memory loads
according to Linux perf profiling. This patch tries to remove the use
of struct priority_req, and make struct rte_mbuf itself carry all the
information. Specifically, it uses the private data of the mbuf to carry
the list information and uses the field udata64 in rte_mbuf to carry
the priority inforation. With this optimization, it removes the need for
struct priority_req as well as the mailbox. Instead, it uses the Ring library
to transfer request packets to the SOL block.