cps: release everything in reverse order in cleanup_cps()
Gatekeeper fails to release the KNI devices when it exits, and
the error message is shown below:

KNI: Fail to release kni device

The root cause of the bug is that Gatekeeper currently removes
the KNI kernel module first and then releases the KNI devices
in cleanup_cps(). The fix is to release everything in reverse order.