lua: improve IPv6 address handling
The use of a reference in str_to_prefix6() in Lua is good enough
for getting an IPv6 address struct from C and passing it to
the lpm6_add() function, but it is not sufficient for keeping
an IPv6 address in memory to later refer to. Therefore,
when we need to inspect the bytes of the result of str_to_prefix6(),
we need an actual struct in6_addr. This patch changes the return
value of str_to_prefix6() to return the actual struct in6_addr.