gt: reply to gateway for non-local traffic
This patch replaces the requirement that Grantor servers had to be
deployed in the same subnet that the protected destination with
the requirement that either Grantor servers are deployed in
the same subnet, or the last hop on the path from a Gatekeeper server
to a Grantor server is a router that can forward
the encapsulated packets to its destinations.

This new requirement supports the deployment environment discussed in
issue #417.