gk: fix checksum issue
The checksum bug discussed in https://github.com/AltraMayor/gatekeeper/issues/417#issuecomment-681891257 is because we have assumed that
the two's complement subtraction, which is currently being used, is the same as
"subtracting complements with borrow" under one's complement as required in
RFC1624. While these two operations often come up with the same result, they
are often not equal too. In addition, the two's complement subtraction is not
endianness preserving!

To solve this issue, we followed the example implmentation of RFC1624 [Eqn. 3]
in Linux kernel:
https://elixir.bootlin.com/linux/latest/source/net/ipv4/netfilter/ipt_ECN.c#L38