gk: fix bug in flow entry deletion
After an expired entry is prefetched to be deleted and
the time at which it is actually removed, the GK block
processes mailbox commands that could remove the entry
or update its expiration time. Therefore, we must check
whether the entry still exists and still is expired
before we remove it.

This patch also adds the flow information to the error
message when deleting a flow entry that is not found.