gk: update num_scan_del correctly
The variable num_scan_del in struct gk_instance must be updated
any time a flow entry is deleted, not just when flow entries
are deleted as a result of flow expiration. Otherwise,
if num_scan_del > 0 (essentially, the flow table is full),
and a dynamic config request to flush the flow table or
delete a FIB entry occurs, num_scan_del will not be updated and
the check in gk_hash_add_flow_entry() will not be accurate.
This will block new entries from being added.