gk: properly initialize new flow entries from policy decisions
When a policy decision was to be applied to a flow entry
that is not already in the flow table, the code was not properly
initializing it, so the flow hash was absent.
See gk/main:update_flow_table() for details.

Instead of simply assigning the flow hash, this patch does
a full initialization of the flow entry, so even if
update_flow_entry() cannot update the entry due to an invalid
policy state, the flow entry is left in a valid state.