launch: clean up master lcore handling
If a stage 3 entry corresponding to the master lcore
is present, it will be run after all other entries,
and therefore it is not necessary to add this entry
to the front of the list.

This patch also enforces that there is at most one
master lcore entry and no other entries in the list
by the time run_master_if_applicable() is called.