gatekeeper: drop privileges by thread
Since the CPS block needs to be run as root in
order to perform certain operations on its
Netlink socket with a userspace routing daemon,
we must individually drop the privileges of
all other blocks by their thread.

Since setuid() and setgid() force all threads
to use the same privileges, we use a syscall
to set the privileges by thread.