net: allow per-block capabilities after user change
If Gatekeeper is configured to run as a non-root
user, then privileges are dropped during startup.
However, some aspects of Gatekeeper (specifically,
being able to use a userspace nl_pid in the CPS
block for interacting with a routing daemon) require
elevated privileges.

In order to reduce privileges of Gatekeeper as a
whole while retaining this functionality, we use
capabilities(7). Each block expresses the capabilities
it needs via the needed_caps() API.

Since Gatekeeper is a multi-threaded system and
DPDK uses a pool of threads, we require calling
prctl(2) for PR_SET_KEEPCAPS for capabilities to
be retained after changing the UID.