gk: only forward from front to back to known neighbors
When a Gatekeeper server is going to forward a packet from
the front interface to a neighbor of the back interface,
Gatekeeper looks up the MAC address of that back neighbor in
the known-neighbor subtable of the FIB.

Gatekeeper has wrongly assumed that any neighbor is always known.
This patch drops packets destined to unknown back neighbors.

This problem was first reported in issue #519.