gk: check the type of a FIB entry before updating it
Currently, only grantor FIB entries can be updated in place.
But the code was not making sure that it was updating a grantor
FIB entry before updating an existing entry.
Thus, the operation would corrupt memory if the passed prefix
were to an existing FIB entry of a different type.