dyc: improve initialization log entries
This patch adds more information to log entries in order to help
users to diagnosis problems.

For example, when the folder `/var/run/gatekeeper/` did not exist,
the log entry said:
"Failed to bind the server socket - (No such file or directory)"
Notice that this message does not include the path.

The new message is as follows:
"Failed to bind the server socket (/var/run/gatekeeper/dyn_cfg.socket), errno=2: No such file or directory"

This patch was motivated by the following issue:
https://github.com/AltraMayor/gatekeeper/issues/546#issuecomment-1111531571