lib/net: boot with bond interfaces
After upgrading Gatekeeper's DPDK, Gatekeeper is not booting with
bond interfaces.  This issue is identified with log entries similar
to the one below:
memzone eth_p1_q0_rx_ring does not justify the requested attributes

The example log entry above and others may differ on the name of
the queue that triggers the issue. In the example above,
the name of the queue is 'eth_p1_q0_rx_ring'.