log: replace rte_cpu_to_be_64() with rte_be_to_cpu_64()
When dumping memory, one should use rte_be_to_cpu_64() instead of
rte_cpu_to_be_64() because one is dumping the memory in the same
order the bytes are in memory. That is, the first byte in memory
is the most significant when printed out, the second byte
the second most significant, and so forth; this is big endianess.

In practice, this change is purely semantic since
rte_cpu_to_be_64() and rte_be_to_cpu_64() are identical on
little and big endian hosts.