gk: improve log entries of the FIB
The log entries of the FIB are part of the day-to-day interactions
between network administrators and Gatekeeper servers.
This commit improves these entries to bring them up to
the new log format and adds information to these entries.

This commit also improves the strings of Lua errors associated with
the FIB, reviews some error codes, and adds `const` to
the parameter `struct ip_prefix` of many FIB functions.