gatekeeper: support the zero-length prefix
The old libraries rte_lpm and rte_lpm6 do not support
the zero-length prefix, but the new libraries rte_fib and
rte_fib6 do.  This commit removes the previous restrictions.

This commit impacts the routing table of Gatekeeper servers, and
the LPM tables used on Lua policies.

This commit closes #91.