gatekeeper: support variable sizes of RSS hash keys
The original code assumes all RSS hash keys are
10 * 4 = 40 bytes. While this assumption has held well through
the years, recent NICs have keys that are 13 * 4 = 52 bytes long.

This commit enables front and back interfaces to have different
RSS hash keys of different lengths. However, if an interface is
bonded, all the keys of the ports must have the same length.
Otherwise, there would not be a way to find the RSS hash of
a packet in software since there would be multiple values
depending on where the packet arrived at the server.