cps: review the KNI operations
The KNI operation config_network_if was starting the physical
interfaces before they were ready to do so, which would break
the initialization of the interfaces.  A typical symptom of
this problem is the following log entry:

Main/0: net: cannot set the MTU on the front iface (error 16)

The KNI operations config_promiscusity and config_allmulticast
are ignored instead of disabled to avoid upsetting applications
since applications that can call these operations run
as root or have the capabilities to do so.