gk: rename gk/fib.c to gk/rt.c
The original code in gk/fib.c was meant to only use a FIB library,
but it currently uses a FIB and RIB library. Therefore, not only is
renaming it to gk/rt.c (RT is short for routing) more meaningful,
but it also frees the prefix fib for implementing a FIB library in
Gatekeeper.