gk: loosen verification of gateway RT entries
The routing table (RT) has been requiring gateway addresses not
to be included in the prefix of gateway and grantor RT entries.
This commit removes this requirement for gateway entries but
maintains it for grantor entries.

The motivation is to support default routes.
Without this commit, default routes are denied with log entries
similar to the ones below:

DYC/9: check_gateway_prefix(0.0.0.0/0): gateway address is in prefix, so gateway is not a neighbor
DYC/9: check_gateway_prefix(::/0): gateway address is in prefix, so gateway is not a neighbor