gt: protect the call of lookup_policy()
Although the code already uses lua_pcall() to call the Lua
function lookup_policy(), it does not protect the code that sets
the call frame of lookup_policy().  Therefore, if Lua runs out of
memory (or any other error), the Lua state panics.

Not only does this commit protect the code that sets the call frame
of lookup_policy(), but it also runs Lua's garbage collector and
retries lookup_policy() if the failure was due to a lack of memory.