Add QID library
The QID library allows functional blocks to quickly
fetch consecutive IDs, such as to use as indexes of
an array. The library uses a stack to quickly push
and pop IDs onto and from the set of unused IDs.

The GK block's routing table is also changed to use
the QID library. Instead of linearly searching for a
free FIB entry to use, it pops an ID (index) from the
QID library, which serves as the index of the free
FIB entry.