dpdk: update branch for Gatekeeper v1.2
This new branch for Gatekeeper v1.2 is based on DPDK v21.05.
The motivation for this new branch is to solve the LACP disconnect bug
(see the DPDK commit
https://github.com/DPDK/dpdk/commit/fad80ab3698e7f7d3e9c2a28c371da6a17fbc477
).
DPDK v21.05 is the first DPDK release to include the referred commit.

This commit also adds the package python3-pyelftools as a dependency.